Added Texture Scaling
Added texture scaling that allows for nearest-neighbour style scaling. Native javascript adds in smoothed edges that don't look good for this style of game. This new system allows for textures to be defined and then loaded in an array, and a system that saves them to an object for referencing later.